requirements.txt
flask
requests
python-dotenv
langchain
langchain-chroma
langchain-groq
langchain-huggingface
langchain-community
langchain-text-splitters
chromadb
sentence-transformers
langdetect
torch
unstructured
pypdf
python-docx
openpyxl
networkx